audio: Coverity android S /audio/audio_hal [1/1]
PD#SWPL-99871

Problem:
coverity check error

Solution:
Fix the relevant errors

Verify:
local test pass

+++ audio_hal/audio_hw_utils.h
@@ -69,7 +69,7 @@
     }
 
 #define R_CHECK_PARAM_LEGAL(ret, param, min, max, fmt, ...)                                     \
-    if (param < min || param > max) {                                                           \
+    if ((int)param < min || param > max) {                                                           \
         AM_LOGE("%s:%d is illegal, min:%d, max:%d " fmt, #param, param, min, max, ##__VA_ARGS__);\
         return ret;                                                                             \
     }
